(PC-2690) refacto : move selectors to proper directory
