--- src/components/pages/Bookings/FilterByVenue/selectors/selectNonVirtualVenues.js
@@ -1,8 +1,0 @@
-import { createSelector } from 'reselect'
-
-const selectNonVirtualVenues = createSelector(
-  state => state.data.venues || [],
-  venues => venues.filter(venue => venue.isVirtual === false)
-)
-
-export default selectNonVirtualVenues
