--- src/components/pages/Bookings/selectors/selectDigitalOffers.js
@@ -1,8 +1,0 @@
-import { createSelector } from 'reselect'
-
-const selectDigitalOffers = createSelector(
-  state => state.data.offers || [],
-  offers => offers.filter(offer => offer.isDigital === true)
-)
-
-export default selectDigitalOffers
