--- src/components/pages/Bookings/selectors/selectOffersByVenueId.js
@@ -1,13 +1,0 @@
-import createCachedSelector from 're-reselect'
-
-function mapArgsToCacheKey(state, venueId) {
-  return venueId || ''
-}
-
-const selectOffersByVenueId = createCachedSelector(
-  state => state.data.offers || [],
-  (state, venueId) => venueId,
-  (offers, venueId) => offers.filter(offer => offer.venueId === venueId)
-)(mapArgsToCacheKey)
-
-export default selectOffersByVenueId
