+++ src/selectors/data/venuesSelectors.js
@@ -1,4 +1,5 @@
 import createCachedSelector from 're-reselect'
+import { createSelector } from 'reselect'
 
 export const selectVenues = state => state.data.venues
 
@@ -44,3 +45,8 @@ export const selectVenueById = createCachedSelector(
   (state, venueId) => venueId,
   (venues, venueId) => venues.find(venue => venue.id === venueId)
 )((state, venueId = '') => venueId)
+
+export const selectNonVirtualVenues = createSelector(
+  state => state.data.venues || [],
+  venues => venues.filter(venue => venue.isVirtual === false)
+)
